requirements.txt
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain>=0.3.0
langchain-postgres>=0.0.12
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
psycopg-binary>=3.1.0
psycopg2-binary>=2.9.0
pypdf>=5.0.0